requirements.txt
feature_engine==1.0.2
requests==2.25.1
xgboost==1.3.3
pymongo==3.11.3
numpy==1.19.2
tqdm==4.59.0
SQLAlchemy==1.4.2
toml==0.10.2
pandas==1.2.3
delta==0.4.2
python-dotenv==0.17.0
PyMySQL==1.0.2
pyspark==3.1.1
scikit_learn==0.24.1
